requirement.txt
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
eventlet==0.36.1
mysql-connector-python==9.0.0
pandas==2.2.3
openpyxl==3.1.5
bcrypt==4.2.0
python-dotenv==1.0.1
flask-session==0.8.0
python-decouple==3.8
